Add files via upload
This repository contains the source code and assets for my personal portfolio website. Please note that this version of the website is not responsive, meaning it may not adapt well to different screen sizes or devices. It is recommended to view the website on a desktop or laptop computer for the best experience. The website showcases my skills, projects, and achievements, providing a comprehensive overview of my capabilities and experience. While it may not be optimized for mobile or tablet devices, the design and content are still intended to highlight my work and attract potential clients or employers. Feel free to explore the code and provide feedback,